+++ mcmd/commands/disable.py
@@ -1,6 +1,6 @@
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post, put
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight, io
 from mcmd.molgenis.resources import ensure_resource_exists, ResourceType
