+++ mcmd/commands/enable.py
@@ -1,9 +1,9 @@
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post, put
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight, io
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 from mcmd.molgenis.resources import one_resource_exists, ensure_resource_exists, ResourceType
