+++ mcmd/commands/give.py
@@ -7,10 +7,10 @@ from urllib.parse import urljoin
 
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post_form
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight, io
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 from mcmd.molgenis.principals import ensure_principal_exists, detect_principal_type, PrincipalType
 # =========
 # Arguments
