+++ mcmd/commands/import_.py
@@ -9,11 +9,11 @@ import mcmd.config.config as config
 from mcmd.molgenis import api
 from mcmd.github import client as github
 from mcmd.molgenis.client import post_file, get, post
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
-from mcmd.config.home import get_issues_folder
+from mcmd.core.home import get_issues_folder
 from mcmd.io.io import highlight, io
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 from mcmd.utils.file_helpers import scan_folders_for_files, select_path
 
 # =========
