+++ mcmd/commands/run.py
@@ -1,12 +1,12 @@
 import shlex
 
-from mcmd import argparser as arg_parser
-from mcmd.command import command
+from mcmd.core import argparser as arg_parser
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
-from mcmd.config.home import get_scripts_folder
+from mcmd.core.home import get_scripts_folder
 from mcmd.io.io import bold, dim, io
 from mcmd.io.logging import get_logger
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 
 
 # =========
