+++ mcmd/commands/set.py
@@ -5,10 +5,10 @@ import json
 
 from mcmd.molgenis import api
 from mcmd.molgenis.client import get, put
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight, io
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 
 
 # =========
