+++ mcmd/core/command.py
@@ -1,8 +1,8 @@
-from mcmd import history
+from mcmd.core import history
 from mcmd.io import io
 from mcmd.molgenis import auth
 from mcmd.config import config
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 
 
 def command(func):
