+++ mcmd/core/history.py
@@ -1,8 +1,8 @@
 from collections import deque
 from os import path
 
-from mcmd.config.home import get_history_file
-from mcmd.utils.errors import McmdError
+from mcmd.core.home import get_history_file
+from mcmd.core.errors import McmdError
 
 _USER_HISTORY = get_history_file()
