+++ mcmd/molgenis/auth.py
@@ -11,7 +11,7 @@ from requests import HTTPError
 from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.config import config
-from mcmd.utils.errors import McmdError, MolgenisOfflineError
+from mcmd.core.errors import McmdError, MolgenisOfflineError
 
 _username = None
 _password = None
