+++ mcmd/molgenis/request_handler.py
@@ -7,7 +7,7 @@ Error responses can come back in varying forms which this decorator tries to uni
 import requests
 
 from mcmd.molgenis import auth
-from mcmd.utils.errors import McmdError, MolgenisOfflineError
+from mcmd.core.errors import McmdError, MolgenisOfflineError
 
 
 def request(func):
